# internal/http/swagger/swagger.yaml
basePath: /api/v1
definitions:
  Address:
    properties:
      city:
        example: Mumbai
        type: string
      coordinates:
        $ref: '#/definitions/JSONB'
      country:
        example: India
        type: string
      pincode:
        example: "123456"
        type: string
      state:
        example: Maharashtra
        type: string
      street:
        example: 123 Main St
        type: string
    type: object
  BaseResponse:
    properties:
      data: {}
    type: object
...
  CreateResturentInput:
    properties:
      address:
        $ref: '#/definitions/Address'
      license:
        example: license
        type: string
      name:
        example: Alamgiri Resturent
        type: string
    type: object
...
  DataNotFoundError:
    type: object
  FilterFieldPredicate:
    properties:
      field:
        description: Field represents a column for the entity you are filtering
        example: name
        type: string
      operator:
        allOf:
        - $ref: '#/definitions/FilterOp'
        description: Operator represents the filter operation you'd like to perform
          on the field
        enum:
        - eq
        - neq
        - gt
        - gte
        - lt
        - lte
        - in
        - nin
        - like
        - nlike
        - ilike
        - nilike
        - isnull
        - notnull
        - between
        example: eq
      value:
        description: Value represents the value you'd like to filter by
    type: object
  FilterInput:
    properties:
      fields:
        description: Fields represents the fields you want to filter
        items:
          $ref: '#/definitions/FilterFieldPredicate'
        type: array
      sort_keys:
        description: SortKeys represents the sort keys you want to sort by
        items:
          $ref: '#/definitions/SortKey'
        type: array
    type: object
  FilterOp:
    enum:
    - eq
    - neq
    - gt
    - gte
    - lt
    - lte
    - in
    - nin
    - like
    - nlike
    - ilike
    - nilike
    - isnull
    - notnull
    - between
    type: string
    x-enum-varnames:
    - FilterOpEq
    - FilterOpNeq
    - FilterOpGt
    - FilterOpGte
    - FilterOpLt
    - FilterOpLte
    - FilterOpIn
    - FilterOpNin
    - FilterOpLike
    - FilterOpNlike
    - FilterOpIlike
    - FilterOpNilike
    - FilterOpIsnull
    - FilterOpNotnull
    - FilterOpBetween
  ForbiddenAccessError:
    properties:
      code:
        example: FORBIDDEN_ACCESS
        type: string
      message:
        example: You are forbidden from accessing this resource
        type: string
    type: object
  InvalidRequestError:
    properties:
      message:
        example: invalid request
        type: string
    type: object
  JSONB:
    additionalProperties: true
    type: object
...
  PaginationResponse:
    properties:
      data: {}
      page:
        example: 1
        type: integer
      size:
        example: 10
        type: integer
      total:
        example: 1000
        type: integer
    type: object
...
  Resturent:
    properties:
      address:
        $ref: '#/definitions/Address'
      created_at:
        example: 2022-02-16 15:35:10.535606+05:30
        type: string
      id:
        example: a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11
        type: string
      license:
        example: license
        type: string
      name:
        example: Alamgiri Resturent
        type: string
      updated_at:
        example: 2022-02-16 15:35:10.535606+05:30
        type: string
    type: object
  SortKey:
    properties:
      direction:
        description: Direction represents the direction of the sort
        example: asc
        type: string
      field:
        description: Field represents a column for the entity you are sorting
        example: name
        type: string
    type: object
  SystemError:
    properties:
      code:
        example: INTERNAL_SERVER_ERROR
        type: string
      message:
        example: Oops! Something went wrong. Please try again later
        type: string
    type: object
  UnauthorizedError:
    properties:
      code:
        example: UNAUTHORIZED
        type: string
      message:
        example: You are not authorized to access this resource
        type: string
    type: object
...
  UpdateResturentInput:
    properties:
      address:
        $ref: '#/definitions/Address'
      license:
        example: license
        type: string
      name:
        example: Alamgiri Resturent
        type: string
    type: object
...
host: localhost:7700
info:
  contact:
    email: rizwank431@gmail.com
    name: API Support
    url: https://rizwank123.github.io
  description: Resturnet's set of APIs
  termsOfService: http://example.com/terms/
  title: Resturnet API
  version: "1.0"
paths:
...
  /resturent:
    post:
      consumes:
      - application/json
      description: Create a resturent
      operationId: createResturent
      parameters:
      - description: 'Bearer '
        in: header
        name: Authorization
        required: true
        type: string
      - description: Payload
        in: body
        name: in
        required: true
        schema:
          $ref: '#/definitions/CreateResturentInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/Resturent'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/InvalidRequestError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/UnauthorizedError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ForbiddenAccessError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/SystemError'
      security:
      - JWT: []
      summary: Create a resturent
      tags:
      - Resturent
  /resturent/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a resturent by id
      operationId: deleteResturent
      parameters:
      - description: 'Bearer '
        in: header
        name: Authorization
        required: true
        type: string
      - description: Resturent id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/InvalidRequestError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/UnauthorizedError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ForbiddenAccessError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/DataNotFoundError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/SystemError'
      security:
      - JWT: []
      summary: Delete a resturent
      tags:
      - Resturent
    get:
      consumes:
      - application/json
      description: Find a resturent by id
      operationId: findResturentById
      parameters:
      - description: 'Bearer '
        in: header
        name: Authorization
        required: true
        type: string
      - description: Resturent id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/Resturent'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/InvalidRequestError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/UnauthorizedError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ForbiddenAccessError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/DataNotFoundError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/SystemError'
      security:
      - JWT: []
      summary: Find a resturent by id
      tags:
      - Resturent
    put:
      consumes:
      - application/json
      description: Update a resturent
      operationId: updateResturent
      parameters:
      - description: 'Bearer '
        in: header
        name: Authorization
        required: true
        type: string
      - description: Resturent id
        in: path
        name: id
        required: true
        type: string
      - description: Payload
        in: body
        name: in
        required: true
        schema:
          $ref: '#/definitions/UpdateResturentInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/Resturent'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/InvalidRequestError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/UnauthorizedError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ForbiddenAccessError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/DataNotFoundError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/SystemError'
      security:
      - JWT: []
      summary: Update a resturent
      tags:
      - Resturent
  /resturent/filter:
    post:
      consumes:
      - application/json
      description: Filter resturents by criteria
      operationId: filterResturents
      parameters:
      - description: Filter criteria
        in: body
        name: in
        required: true
        schema:
          $ref: '#/definitions/FilterInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/PaginationResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/Resturent'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/InvalidRequestError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/UnauthorizedError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ForbiddenAccessError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/SystemError'
      summary: Filter resturents
      tags:
      - Resturent
...
schemes:
- http
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"